gccrs: Rename header rust-hir-trait-ref.h to rust-hir-trait-reference.h
TODO


gcc/rust/ChangeLog:

	* typecheck/rust-hir-trait-ref.h: Moved to...
	* typecheck/rust-hir-trait-reference.h: ...here.
	* typecheck/rust-hir-trait-resolve.cc:
	* typecheck/rust-hir-trait-resolve.h (RUST_HIR_TRAIT_RESOLVE_H):
	* typecheck/rust-hir-type-check.h:
	* typecheck/rust-tyty.cc:
