+++ gcc/rust/typecheck/rust-hir-trait-reference.h
